Added functionality to only get the VPN IP addresses of parent or child tasks
This is a common use case: a child task wants to communicate to the parent or a parent to their children. Previously, the algorithm developer would then be presented with a dictionary from which they had to determine which tasks were parents/children/same level. Now, this can be done by a simple function call