Add 'smtp.email_from' option and use it as sender email
sender email  =  email 'from' (header)

The 'username' used for authenticating with the SMTP server is not
necessarily the address vantage6 server admins want to send emails from
to their users. In fact, they might rather keep that information (smtp
username) not public.

This approach of explicitly setting 'email from' address is similar to
what for example GitLab [0] or Gitea [1] do.

Support email address is no longer used in headers. It's simply an email
address for users to be aware of. It's used in email body.

[0]: https://docs.gitlab.com/omnibus/settings/smtp.html
[1]: https://docs.gitea.com/administration/email-setup#using-smtp