Add function documentation & refactor handle_password_recovery
Goal of moving the reading of settings to top is to minimize chances of
leaking non-constant time behavior under heavy load. This however has
not been tested.

We no longer check for the validity of arguments as we assume this has
been done before (Schema validation). However, function has been
prefixed with '_' to (lightly) hint it's not meant for general use.

This addresses review comments.