Fix: allow for setting node container name in config
By default the container name is taken from the config file name, which
means you cound't start multiple nodes that have the same filename for
their config file, like the ones under `dev/`.

We could also simply change the name of the config file for the multiple
nodes, but then we would also need to pass it the name of the config
file to the container node, as the default is `config.yaml`.