When creating a task, multiple dropdown selection property was not changed if function type changed
E.g. when first selection a partial function, you would get a multi-select dropdown for the organizations to create the task for. However, when you then changed your mind and selected a central function, the dropdown did not change to only allow you to select a single organization

This turned out to be so because the multiple property of mat-select cannot be changed after initialization