Handle error when user is not registered in algorithm store
When going to algorithm store management pages, and user is not registered in a store, checking their permission gives a 403. This was not caught properly leading to an error message in the UI - while everything was functional as they are not allowed to change anything but maybe see algorithms. Now the error is caught