Indefinitely try to obtain new node token if the token expired
In v4, trying to get a new token was cancelled at some point if the refresh token had expired, because there was no way of getting a new token anymore. But now that we are using keycloak service account, there is no refresh token, so we can just keep trying to get a token. This makes nodes more stable to reauthenticate if we have a server that goes offline for a few hours or more